dashboard: Fix implicit `float` to `int` conversion warning with PHP 8.1

+++ dashboard/dashboard.php
@@ -21,14 +21,14 @@
 			</tr>
 		</thead>
 		<tbody>
-		<?php foreach (array_reverse(getDonations()) as $donation) { ?>
+<?php foreach (array_reverse(getDonations()) as $donation) { ?>
 			<tr>
 				<td><?= $donation['username'] ?></td>
 				<td><?= formatCurrency($donation['amount'], $donation['currency']) ?></td>
-				<td><?= date('d.m.Y H:i:s', $donation['time'] / 1000) ?></td>
+				<td><?= date('d.m.Y H:i:s', (int)($donation['time'] / 1000)) ?></td>
 				<td><?= $donation['message'] ?></td>
 			</tr>
-		<?php } ?>
+<?php } ?>
 		</tbody>
 	</table>
 </div>
